Added "undef" as an alternative to 0 when the substitution could not be performed
param_exists() returns 1 if the parameter is present and can be
substituted, 0 if it is absent, and undef if the substitution failed.
param_is_defined() returns 1 if the paramter is present and can be
substituted to a defined value, undef if the substitution fails, 0
otherwise.
param() returns the value if param_exists() returned true, undef otherwise.
param_required() is like param() but dies instead of returning undef.