bugfix for the infinite loop on "unreachable" workers
Queen::check_for_dead_workers() calls Valley::status_of_all_our_workers_by_meadow_signature()
and checks there not only the status of the workers, but also which meadows
are available.  In 2.4, this used to contain empty hashes for reachable
meadows with no workers, but this behaviour was broken on master when I
refactored parts of Valley and Meadow.  In case there was no running
workers, there wouldn't even be an entry for the meadow, and Queen was
considering the meadow to be unreachable.  If at the same time a worker is
killed by LSF on this meadow, beekeeper wasn't able to understand it could
bacct it.