Cache the total hive load in HivePipeline
We noticed that beekeeper was sometimes very slow doing anything on the
database. This was tracked down to running a query over and over (from
RoleAdaptor::get_hive_current_load()).
This commit introduces two methods in HivePipeline: one to fetch + cache the
value, and one to invalidate the cache. Other modules have been adapted to
invalidate the cache only when needed, resulting in an optimal use of the
database.