Fixed the "topup" mode of MySQLTransfer
The "topup" mode was not error-safe because it the first run failed, the
second would almost certainly fail because of duplicated rows.
The solution is to add the --insertignore flag and to tweak the post-copy
checks. Even though the mysqldump options are now the same in "topup" and
"insertignore" modes, "topup" still means that all the source rows must be
in the target table (even though some could already be present when the
copy starts). "insertignore" means that we know that some rows could be
missing.