Use File::Glob to get the list of filenames (#21)
* Use File::Glob to get the list of filenames

Previously backtick ls command was used to fetch the list of sql patches filenames.
By using backtick it is required to fork a new process. This would potentially be
a problem for a running process that uses large amount of memory. If there is not
enough spare memory left in the system, the fork operation could be aborted because
of "cannot allocate memory" error.

Fixes ENSEMBL-4823.