beekeeper.pl::big_red_button: exclude self from the "just blocked" list
A beekeeper executing the shutdown cannot loop so it isn't strictly
speaking necessary to block it, on the other hand it might perhaps come
in handy (e.g. for diagnostic purposes) to have all beekeepers involved
in a shutdown get is_blocked set to 1. Therefore, the actual blocking
still includes self but we filter that one entry out of the report.