docs/xhive/code_doc: avoid using basestring under Python3
One way of doing this would be to refactor get_command() so that it
never outputs bytestrings, that said it feels like putting too much
effort into maintaining compatibility with a Python version that will
reach end of life within 2 years. Just choose the right string type
depending on the major version of Python, that way all that will have to
be done once we have decided to drop python2 support is to revert to
unconditionally checking _command_ against a single type.