docs/xhive/analysis_diagram: decode subprocess output
Our documentation builder calls external scripts to produce Graphviz
dot code representing pipelines.

On the one hand, under both Python2 and Python3 the subprocess output
is a bytestring. On the other, some internals Sphinx Graphviz module
expects dot code to have string representation, which is NOT the same
as bytestrings under Python3. As a result, when running under Python3
the Graphviz module fails to render dot code due to being unable to
concatenate a string with a bytestring.

Now subprocess output is decoded to string, using the default encoding
of the given Python version. This makes the Graphviz module happy.