Set of minimal dependencies
pom.xml now only lists the direct dependencies (since maven takes care of
the indirect ones). This solves two issues that pom.xml had:
- different versions of slf4j required
- duplicated entries

For instance, it fixes this warning,
```
[WARNING]
[WARNING] Some problems were encountered while building the effective model for org.ensembl.hive:eHive:jar:1.0
[WARNING] 'dependencies.dependency.(groupId:artifactId:type:classifier)' must be unique: org.slf4j:slf4j-api:jar -> version 1.7.5 vs 1.7.12 @ line 65, column 17
[WARNING]
[WARNING] It is highly recommended to fix these problems because they threaten the stability of your build.
[WARNING]
[WARNING] For this reason, future Maven versions might no longer support building such malformed projects.
[WARNING]
```

I have also changed the versions to be the same ones as in the original
build.xml. The PR is about changing the build system, not the list /
versions of the dependencies.