Travis: Add mysql and postgresql to the list of services
Ubuntu Xenial, which is now being rolled out as the default Linux distro
for Travis builds following the end of life of Ubuntu Trusty, does not
start services such as MySQL or PostgreSQL unless told to. We use the
default distro in our builds so we are likely to be affected.